visualizations/overview/backend/requirements/base.txt
Django==3.2.13
djangorestframework==3.13.1
whitenoise==5.2.0
django-cors-headers
altair==4.2.0
altair-saver==0.5.0
numpy
mysqlclient